initFirebase.js added
This file is responsible for initializing the firebase app and allowing all files to use firebase features
