+++ config/initFireBase.js
@@ -1,0 +1,33 @@
+require("dotenv").config();
+const admin = require("firebase-admin");
+
+ const serviceAccount = {
+   "type": "service_account",
+     "project_id": "ian-14014",
+     "private_key_id": process.env.FIRE_SERVICE_PRIVATE_KEY_ID,
+     "private_key": process.env.FIRE_SERVICE_PRIVATE_KEY.replace(/\\n/g, '\n'),
+     "client_email": process.env.CLIENT_EMAIL ,
+     "client_id": process.env.CLIENT_ID,
+     "auth_uri": process.env.AUTH_URI,
+     "token_uri": process.env.TOKEN_URI,
+     "auth_provider_x509_cert_url": process.env.AUTH_PROVIDER_X509_CERT_URL,
+     "client_x509_cert_url": process.env.CLIENT_X509_CERT_URL,
+     "universe_domain": "googleapis.com"
+ }
+
+ admin.initializeApp({
+   credential: admin.credential.cert(serviceAccount),
+ });
+ module.exports = {admin};
+const firebase = require("firebase/app");
+   const firebaseConfig = {
+   apiKey: process.env.FIREBASE_API_KEY,
+   authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+   projectId: process.env.FIREBASE_PROJECT_ID,
+   storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+   messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID,
+   appId: process.env.FIREBASE_APP_ID,
+   measurementId: process.env.FIREBASE_MEASUREMENT_ID
+ };
+
+firebase.initializeApp(firebaseConfig);
