+++ views/LoginUser.ejs
@@ -33,7 +33,11 @@
                 <button data-mdb-button-init data-mdb-ripple-init class="btn btn-info btn-lg btn-block" >Login</button>
               </div>
               <hr>
-              <button  ><a href='/auth'>Login With Google</a></button>
+              <div id="google-btn-container" class="google-btn-container">
+                <span class="google-text">
+                  Continue with Google
+                </span>
+              </div>
               <p>Don't have an account? <a href="/CreateAccount" class="link-info">Register here</a></p>
         
             </form>
@@ -49,6 +53,51 @@
     </div>
   </section>
   <script src ="/js/LoginAccount.js"></script>
+  <!-- <script type="module">
+      // Import Firebase modules
+      import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.14.0/firebase-app.js';
+      import { getAuth, GoogleAuthProvider, signInWithPopup } from 'https://www.gstatic.com/firebasejs/9.14.0/firebase-auth.js';
 
+      // Your Firebase config
+      const firebaseConfig = {
+          
+      };
+
+      // Initialize Firebase
+      const app = initializeApp(firebaseConfig);
+      const auth = getAuth(app);
+
+      document.addEventListener('DOMContentLoaded', () => {
+          document.getElementById('google-btn-container').addEventListener('click', () => {
+              const provider = new GoogleAuthProvider();
+              signInWithPopup(auth, provider)
+                  .then((result) => {
+                      const token = result._tokenResponse.idToken; // Use this token
+                      fetch('/googleLogin', {
+                          method: 'POST',
+                          headers: {
+                              'Content-Type': 'application/json',
+                          },
+                          body: JSON.stringify({ token }),
+                      })
+                      .then(response => response.json())
+                      .then(data => {
+                          // if (data.message === 'Successfully authenticated') {
+                          //     // Redirect on the client side
+                          //     window.location.href = `/home?email=${encodeURIComponent(data.email)}`;
+                          // } else {
+                          //     console.error('Authentication failed:', data.error);
+                          // }
+                      })
+                      .catch((error) => {
+                          console.error('Error:', error);
+                      });
+                  })
+                  .catch((error) => {
+                      console.error('Error during sign in:', error);
+                  });
+          });
+      });
+  </script> -->
 </body>
 </html>
